==> 00-clone_projects.ipynb <==
This notebook helps clone and update the projects

import pandas as pd

import os

projects = pd.read_pickle('projects.zip')

# Clone Projects

projects['name'] = projects.index

clone_all = False
for i,project in enumerate(projects.itertuples()):
    project_path = os.path.abspath(os.path.join('..','..','..','repos2',project.name))
    git_path = os.path.abspath(os.path.join(project_path,'.git'))
    if clone_all or not os.path.exists(git_path):
        print(f"mkdir -p {project.name}; git clone git://github.com/{project.name} {project.name}")

# Update Projects

for i,project in enumerate(projects.itertuples()):
    project_path = os.path.abspath(os.path.join('..','..','..','repos2',project.name))
    git_path = os.path.abspath(os.path.join(project_path,'.git'))
    if os.path.exists(git_path):
        print(f"cd {project_path}; git pull;")

for i,project in enumerate(projects.itertuples()):
    print(project.name)
