==> 01-project_characterization.ipynb <==
pd.options.display.max_rows = 500

projects = pd.read_pickle('projects.zip')
projects.head(5)

releases = pd.read_csv('releases.csv')
releases = releases.set_index(['project', 'name'])
releases.head(5)

Number of projects:

releases.groupby(level=0).count().shape

Stars

prj_lang_grp = projects.groupby('lang')
rl_lang_grp = releases.groupby('lang')
summary = pd.DataFrame(index=prj_lang_grp.count().index)
summary['mean stars'] = prj_lang_grp.mean()['stars']
summary['std stars'] = prj_lang_grp.std()['stars']
summary['total stars'] = prj_lang_grp.sum()['stars']
summary['total commits'] = rl_lang_grp.sum()['commits']
summary['total releases'] = releases.reset_index('name').groupby('lang').count()['name']
summary


Commits

pd.DataFrame(summary.sum()).T

x = pd.DataFrame(summary.mean()).T
x.index = ['sum']
x
summary.append(x)

releases.reset_index('name').groupby('lang').count()['name'].sum()
