==> 03-inspect.ipynb <==
import pandas as pd
import numpy as np
%matplotlib widget
import matplotlib.pyplot as plt
from scipy import stats
import re
pd.options.display.max_rows = 500

releases = pd.read_pickle('releases.zip')

releases.loc['ytdl-org/youtube-dl']

releases[(releases.range_recall == 1) & (releases.range_precision < 1)][['range_recall', 'project', 'name', 'base_releases', 'range_base_releases']]

releases[releases.range_recall < 1][['range_recall', 'project', 'name', 'base_releases', 'range_base_releases']]
