==> 03-samples.ipynb <==
import pandas as pd

releases = pd.read_csv('releases.csv')

releases = releases.set_index(['project', 'name'])

releases[(releases['commits'] < 10) & (releases['time_precision'] < 1) & (releases['time_recall'] < 1) & (releases['range_precision'] == 1) & (releases['commits'] > releases['time_commits'])]

releases[(releases['commits'] < 10) & (releases['range_precision'] < 1)]

releases[(releases['commits'] < 10) & (releases['range_recall'] < 1)]
